マルチブート整備

+++ prog/src/multiboot/comm/mbp.h
@@ -153,7 +153,7 @@ MBPChildState;
 /* variable */
 
 /* 親機初期化 */
-void    MBP_Init(u32 ggid, u16 tgid);
+void    MBP_Init(u32 ggid, u32 tgid);
 void    MBP_Start(const MBGameRegistry *gameInfo, u16 channel);
 
 /* 1 フレームごとの親機メイン処理 */
